feat: implement authentication flow
